Replace SymSpell with URIK binary dictionaries (#797)
- Removes SymSpell dependency entirely
- Word coverage expanded from ~25-35k per language to ~100k+

+++ app/src/main/java/com/urik/keyboard/ui/keyboard/components/ZipfCheck.kt
@@ -32,7 +32,7 @@ constructor(private val spellCheckManager: SpellCheckManager) {
         geometricAnalysis: PathGeometryAnalyzer.GeometricAnalysis,
         keyPositions: Map<Char, PointF>,
         bigramPredictions: Set<String>,
-        wordFrequencyMap: Map<String, Int>,
+        wordFrequencyMap: Map<String, Long>,
         pathSize: Int
     ): ArbitrationResult {
         if (scoredCandidates.isEmpty()) {
@@ -274,7 +274,7 @@ constructor(private val spellCheckManager: SpellCheckManager) {
 
     private fun enrichWithPrefixCompletions(
         candidates: List<WordCandidate>,
-        wordFrequencyMap: Map<String, Int>,
+        wordFrequencyMap: Map<String, Long>,
         pathSize: Int
     ): List<WordCandidate> {
         if (candidates.isEmpty()) return candidates
